Extract raw config files
