+++ platform/config/index.ts
@@ -1,6 +1,7 @@
 export { ConfigService } from './ConfigService';
-export { RawConfigService, RawConfig } from './RawConfigService';
+export { RawConfigService } from './RawConfigService';
+export { RawConfig } from './RawConfig';
 /** @internal */
-export { ObjectToRawConfigAdapter } from './RawConfigService';
+export { ObjectToRawConfigAdapter } from './ObjectToRawConfigAdapter';
 export { Env } from './Env';
 export { ConfigWithSchema } from './ConfigWithSchema';
